Added unwrap() to match update to RWLock.
Now that the write() and read() functions for locks return a LockResult
to notify users when a thread has panic'ed while holding the lock the
LockResult needs to be unwrapped to match previous behavior.

There are likely some cases where the poisoned guard case can be handled
more gracefully than a panic but this at least lets the project compile.