Rename Context.uri to Context.uri_path
BREAKING CHANGE

Context.uri was misleading as it represented the path of the URI and not
the entire URI. The `Uri` type was renamed to `UriPath` and
`Context.uri` was renamed to `Context.uri_path`.

This change modifies the public interface of rustful. Anyone depending
on `Context.uri` will encounter compiler errors.

Fixes #115